[REFACTOR] dto 이름 변경 및 필드 camelCase로 수정 #357
