+++ src/main/java/FIS/iLUVit/dto/post/PostResponse.java
@@ -1,90 +1,65 @@
 package FIS.iLUVit.dto.post;
 
-import FIS.iLUVit.dto.comment.CommentDto;
 import FIS.iLUVit.domain.Post;
+import com.querydsl.core.annotations.QueryProjection;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Lob;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 public class PostResponse {
 
-    private Long id; // 게시글 아이디
-    private Long writer_id;
-    private String nickname;
+    private Long postId;
+    private String title;
+    private String content;
     private Boolean anonymous;
+    private String nickname; // 작성자 닉네임
+    private Long userId; // 작성자 id
+    private String profileImage; // 작성자 프로필 이미지
+    private int commentCnt;
+    private int heartCnt;
+    private int imgCnt;
+    private String previewImage;
+    private String board_name;
+    private Long boardId;
     private LocalDate date;
     private LocalTime time;
-    private String title;
-    @Lob
-    private String content;
-
-    private String profileImage;
-    private List<String> images;
-
-    private Integer imgCnt;
-    private Integer heartCnt;
-
-    private String boardName;
+    private Long centerId;
     private String centerName;
 
-    private List<CommentDto> comments;
-    private Integer commentCnt;
-
-    private Long boardId;
-    private Long centerId;
-
-    private Boolean canDelete;
-
-    public PostResponse(Post post, List<String> encodedImages, String encodedProfileImage, Long userId) {
-        this.id = post.getId();
-        if (post.getUser() != null) {
-            if (Objects.equals(post.getUser().getId(), userId)) {
-                this.canDelete = true;
-            } else {
-                this.canDelete = false;
-            }
-            if (post.getAnonymous()) {
-                this.nickname = "익명";
-            } else {
-                this.writer_id = post.getUser().getId();
-                this.nickname = post.getUser().getNickName();
-            }
-        }
-        this.anonymous = post.getAnonymous();
-        this.date = post.getDate();
-        this.time = post.getTime();
+    @QueryProjection
+    public PostResponse(Post post) {
+        this.postId = post.getId();
         this.title = post.getTitle();
         this.content = post.getContent();
-
-        this.profileImage = encodedProfileImage;
-        this.images = encodedImages;
-
-        this.imgCnt = encodedImages.size();
-        this.heartCnt = post.getHeartCnt();
-
-        if (post.getBoard() != null) {
-            this.boardId = post.getBoard().getId();
-            this.boardName = post.getBoard().getName();
-            if (post.getBoard().getCenter() != null) {
-                this.centerId = post.getBoard().getCenter().getId();
-                this.centerName = post.getBoard().getCenter().getName();
-            }
-        }
-
-        this.comments = post.getComments().stream()
-                .filter(c -> c.getParentComment() == null)
-                .map(c -> new CommentDto(c, userId)).collect(Collectors.toList());
+        this.anonymous = post.getAnonymous();
+        this.nickname = post.getUser().getNickName();
+        this.userId = post.getUser().getId();
+        this.profileImage = post.getUser().getProfileImagePath();
         this.commentCnt = post.getCommentCnt();
+        this.heartCnt = post.getHeartCnt();
+        this.imgCnt = post.getImgCnt();
+        this.board_name = post.getBoard().getName();
+        this.boardId = post.getBoard().getId();
+        this.date = post.getDate();
+        this.time = post.getTime();
+        this.previewImage = post.getInfoImagePath();
 
+        if (post.getBoard().getCenter() != null) {
+            this.centerId = post.getBoard().getCenter().getId();
+            this.centerName = post.getBoard().getCenter().getName();
+        } else {
+            this.centerName = "모두의 이야기";
+        }
+    }
+
+    public void updatePreviewImage(List<String> encodedInfoImage) {
+        this.previewImage = encodedInfoImage.isEmpty() ? null : encodedInfoImage.get(0);
     }
 }
