+++ src/main/java/FIS/iLUVit/service/PostHeartService.java
@@ -19,9 +19,7 @@ public class PostHeartService {
     private final PostHeartRepository postHeartRepository;
 
     /**
-     작성자: 이창윤
-     작성시간: 2022/06/27 1:40 PM
-     내용: 게시글에 이미 좋아요 눌렀는지 검증 후 저장
+     * 게시글 좋아요 등록
      */
     public Long savePostHeart(Long userId, Long postId) {
         if (userId == null) {
@@ -41,6 +39,9 @@ public class PostHeartService {
         return postHeartRepository.save(postHeart).getId();
     }
 
+    /**
+     * 게시글 좋아요 취소 ( 기존에 좋아요 눌렀던 상태여야 취소 가능 )
+     */
     public void deletePostHeart(Long userId, Long postId){
         if (userId == null) {
             throw new UserException(UserErrorResult.NOT_VALID_TOKEN);
